requirements.txt
django==2.0.6
django-allauth
django-grappelli
djangorestframework
psycopg2-binary
django-oauth-toolkit
django-cors-middleware
django-extensions
django-log-request-id
django-rest-swagger
django-cors-headers
django-profile-middleware
django-silk
djutil
XlsxWriter
python-memcached
python-dateutil
django-rest-auth
django-allauth
djangorestframework-jwt